Refactor: baseResponse타입
